extract VirtualRows component from InfiniteScroll to separate scroll-driven re-renders from data merge logic, increase overscan from 10 to 50 rows for smoother fast scrolling, add willChange transform hint to virtualized container, replace inline scroll handler with useRef guard pattern to avoid re-subscriptions, consolidate scroll reset triggers into single resetKey prop, and switch TableRows keys from index to stable IDs (positionId, txId/txHash, limitOrderId)
